Improve VersionComponentTest with traits.
- Separate data providers to a trait.
- Use ``TestConstructorHelper``.
