+++ tests/VersionComponentTest.php
@@ -34,51 +34,12 @@ use \InvalidArgumentException;
 class VersionComponentTest extends TestCase
 {
     use ExporterPlugin;
+    use VersionComponentTestProvider;
+    use TestConstructorHelper;
 
-    /**
-     * @coverage VersionComponent::__construct
-     * @param integer $int [description]
-     * @param string $str [description]
-     *
-     * @return void
-     * @dataProvider constructorArgumentsProvider
-     */
-    public function testCreatesNewInstances($int, $str)
+    public function getTargetClassName()
     {
-        $obj = new VersionComponent($int, $str);
-        $this->assertInstanceOf(VersionComponent::class, $obj);
-    }
-
-    public function constructorArgumentsProvider()
-    {
-        return [
-            'null values'       => [null, null],
-            'Only integer part' => [1, null],
-            'Only string part'  => [null, '-alpha'],
-            'All arguments'     => [5, '-beta'],
-            'Git describe'      => [19, '-g7575872'],
-        ];
-    }
-
-    /**
-     * @coverage VersionComponent::__construct
-     * @expectedException InvalidArgumentException
-     * @dataProvider badConstructorArgumentsProvider
-     */
-    public function testInformsWhenErrorOccursOnCreatingNewInstances($major, $minor, $build = null, $rev = null)
-    {
-        $obj = new VersionComponent($major, $minor, $build, $rev);
-    }
-
-    public function badConstructorArgumentsProvider()
-    {
-        return [
-            'Negative integer part'        => [-1, null],
-            'Invalid string value part'    => [0, 'erróneo'],
-            'Invalid type (float) for string part'  => [0, 23.912],
-            'Invalid type (object) for string part'  => [0, new \stdClass],
-            'Invalid type (array) for string part'  => [0, ['no']],
-        ];
+        return VersionComponent::class;
     }
 
     /**
@@ -131,7 +92,7 @@ class VersionComponentTest extends TestCase
      * @coverage VersionComponent::equals
      * @coverage VersionComponent::compareTo
      * @dataProvider compareToProvider
-     * @depends testCreatesNewInstances
+     * @[depends] testConstructor
      * @[depends] testPerformsConversionFromString
      */
     public function testCanCompareWithOtherObjects($expected, VersionComponent $left, $right)
@@ -170,48 +131,13 @@ class VersionComponentTest extends TestCase
         }
     }
 
-    public function compareToProvider()
-    {
-        $v = new VersionComponent(1, '-alpha');
-        $obj = new \stdClass();
-        $obj->intValue = 1;
-        $obj->stringValue = '-alpha';
-
-
-        $args = [
-            'Equals by reference' => [0, $v, $v],
-            'Equals by value'     => [
-                0,
-                new VersionComponent(1, '-alpha'),
-                VersionComponent::parse('1-alpha')
-            ],
-            'VersionComponent: >' => [
-                1,
-                new VersionComponent(1, '-beta'),
-                VersionComponent::parse('1-alpha')
-            ],
-            'VersionComponent: <' => [
-                -1,
-                new VersionComponent(1, '-alpha'),
-                VersionComponent::parse('1-beta')
-            ],
-            'VersionComponent | stdClass: null' => [
-                null,
-                $v,
-                $obj
-            ],
-        ];
-
-        return $args;
-    }
-
     /**
      * @coverage VersionComponent::equals
      * @coverage VersionComponent::compareTo
      * @coverage Object::compare
      * @depends testCanCompareWithOtherObjects
      * @depends NelsonMartell\Test\ObjectTest::testProvidesSortingInArrays
-     * @dataProvider comparationObjectsProvider
+     * @dataProvider compareProvider
      */
     public function testCanBeSortedInArrays(array $expected)
     {
@@ -231,39 +157,5 @@ class VersionComponentTest extends TestCase
         );
 
         $this->assertEquals($expected, $actual, $message);
-    }
-
-    public function comparationObjectsProvider()
-    {
-        return [
-            'VersionComponent[]' => [[
-                VersionComponent::parse("0-4-g"),
-                VersionComponent::parse("1-4-g"),
-                VersionComponent::parse("2-3-g"),
-                VersionComponent::parse("2-3-g726356"),
-                VersionComponent::parse("2-4-g"),
-                VersionComponent::parse("4-3-g"),
-                VersionComponent::parse("4-3-gsh4hajk7"),
-                VersionComponent::parse("4-3-gsh4hbjk7"),
-                VersionComponent::parse("11-4-g"),
-            ]],
-            'VersionComponent[] + integer[]' => [[
-                1,
-                new VersionComponent(2, '-alpha'),
-            ]],
-            'VersionComponent[] + string[]'  => [[
-                new VersionComponent(1, '-alpha'),
-                '1-beta',
-            ]],
-            'VersionComponent[] + string[] (non parseable)'  => [[
-                '----------',
-                new VersionComponent(),
-            ]],
-            'VersionComponent[] + array[]'   => [[
-                [],
-                [0, 1, 0],
-                new VersionComponent(1, '-alpha'),
-            ]],
-        ];
     }
 }
